requirements.txt
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
lxml==5.4.0
packaging==25.0
pillow==11.2.1
pip3-autoremove==1.2.2
pipdeptree==2.26.1
requests==2.32.4
setuptools==80.9.0
soupsieve==2.7
typing_extensions==4.14.0
urllib3==2.5.0
